renamed active auctions component file to auctions since it now handles all general auction related features
